OvmfPkg: Fix VS2005 compiler warnings
v2:
 * Incorporate changes suggested by Laszlo

Contributed-under: TianoCore Contribution Agreement 1.0
Cc: Laszlo Ersek <lersek@redhat.com>

+++ OvmfPkg/Library/LoadLinuxLib/Linux.c
@@ -384,8 +384,8 @@ SetupLinuxMemmap (
 #ifdef MDE_CPU_IA32
   Efi->efi_loader_signature = SIGNATURE_32 ('E', 'L', '3', '2');
 #else
-  Efi->efi_systab_hi = ((UINT64)(UINTN) gST) >> 32;
-  Efi->efi_memmap_hi = ((UINT64)(UINTN) MemoryMapPtr) >> 32;
+  Efi->efi_systab_hi = (UINT32) (((UINT64)(UINTN) gST) >> 32);
+  Efi->efi_memmap_hi = (UINT32) (((UINT64)(UINTN) MemoryMapPtr) >> 32);
   Efi->efi_loader_signature = SIGNATURE_32 ('E', 'L', '6', '4');
 #endif
